COMP: quiet warnings
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=4937
git-svn-id: http://svn.slicer.org/Slicer4/trunk@4937 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Modules/EMSegment/MRML/vtkMRMLEMSTreeNode.cxx
@@ -225,7 +225,7 @@ const char*
 vtkMRMLEMSTreeNode::
 GetNthChildNodeID(int n)
 {
-  if (n >= 0 && n < this->ChildNodeIDs.size())
+  if (n >= 0 && n < static_cast<int>(this->ChildNodeIDs.size()))
     {
     return this->ChildNodeIDs[n].c_str();
     }
@@ -241,7 +241,7 @@ vtkMRMLEMSTreeNode::
 GetChildIndexByMRMLID(const char* childID)
 {
   std::string searchID(childID);
-  for (int i = 0; i < this->ChildNodeIDs.size(); ++i)
+  for (int i = 0; i < static_cast<int>(this->ChildNodeIDs.size()); ++i)
     {
     if (this->ChildNodeIDs[i] == searchID)
       {
